Support comments, support wide characters.
In one of the specimen a character constant has two characters in it.
No explanation in the Users' Reference, but in the Tutorial it is
explained that (at their site) up to 4 characters can fit into 1 36-bit word.
In the Users' Reference only up to 2 characters at a time are used,
which can fit in a 16-bit word.